requirements.txt
Flask
mysql-connector-python
flask-swagger-ui
python-dotenv
flask-cors
paho-mqtt
flask-socketio 
gevent 
gevent-websocket 
eventlet
asyncio
websockets
thread
FastAPI
python-multipart
uvicorn
WebSocket
aiofiles
pytz
